git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@1279 45b28200-7c74-4d1a-96bc-342942c5875a

+++ src/fr/ird/osmose/UniformPlankton.java
@@ -68,7 +68,7 @@ public class UniformPlankton extends Plankton {
     /**
      * The biomass, in tonne, in a cell of the model.
      */
-    private float biomass;
+    private double biomass;
 
 //////////////
 // Constructor
@@ -92,7 +92,7 @@ public class UniformPlankton extends Plankton {
     @Override
     public void init() {
         super.init();
-        biomass = (float) (getConfiguration().getFloat("plankton.biomass.total.plk" + getIndex()) / getGrid().getNOceanCell());
+        biomass = getConfiguration().getFloat("plankton.biomass.total.plk" + getIndex()) / getGrid().getNOceanCell();
     }
 
     /**
@@ -104,7 +104,7 @@ public class UniformPlankton extends Plankton {
      * @return the biomass of plankton in this {@code cell}
      */
     @Override
-    public float getBiomass(Cell cell) {
+    public double getBiomass(Cell cell) {
         return biomass;
     }
 }
